+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## v1.0.0 
 - Class representation (`__repr__`) added on NepaliTimeZone
 - Refactored converter
+- Added locations feature
+- Added Phone number parse feature
 
 ## v0.5.6 - (July 7, 2022)
 - Test for automatic package publish